+++ CHANGES.md
@@ -7,6 +7,9 @@ but this file may sometimes contain later improvements (e.g. typo fixes).
 
 ## next (not yet released)
 
+- BREAKING CHANGE:
+  m3api-botpassword now requires at least Node.js 18.2.0,
+  up from Node 12.22.0 or 14.17.0 previously.
 - `login()` now detects if logging in failed,
   and throws a `LoginError` in that case.
   (This new class is also exported,